Merge in SSBD2024/ssbd202401 from feature/SSBD202401-255-resetowanie-hasla-dla-zablokowanego-konta to master

* commit '80203bece46226aa9572bc024188b1672574274a':
  checkstyle
  Verification before sending mail added

+++ src/main/java/pl/lodz/p/it/ssbd2024/ssbd01/mok/service/AccountService.java
@@ -240,7 +240,9 @@ public class AccountService {
         if (account.isEmpty()) {
             return;
         }
-
+        if (!account.get().getActive() || !account.get().getVerified() || !account.get().isAccountNonLocked()) {
+            return;
+        }
         resetCredentialRepository.deleteByAccount(account.get());
         resetCredentialRepository.flush();
         changeMyPasswordRepository.deleteByAccount(account.get());
